isolating toasts to controll animations when toast appears and disappears
